+++ CHANGELOG.md
@@ -5,6 +5,8 @@ HEAD
 ----
 
 * Fixed invalid conversion error in `operator|(JsonVariant, char*)` (issue #1432)
+* Changed the default value of `ARDUINOJSON_ENABLE_PROGMEM` (issue #1433).
+  It now checks that the `pgm_read_XXX` macros are defined before enabling `PROGMEM`.
 
 v6.17.1 (2020-11-07)
 -------